requirements.txt
flask
flask-socketio
pyserial
